Add missing helper functions
