+++ SpotifyYoutubeSync/controllers/basic_controllers.py
@@ -4,7 +4,7 @@ from urllib.parse import urlencode
 import base64
 import json
 from SpotifyYoutubeSync import app
-
+from ..helpers.helpers import check_enable_sync_button, get_video_names_from_playlist, spotify_search_for_songs, add_tracks_to_spotify_playlist
 
 @app.route("/")
 def hello_world():
@@ -16,11 +16,15 @@ def app_home():
     return render_template("home.html", enable_sync_button=enable_sync_button)
 
 
-def check_enable_sync_button():
-    res = False
-    if session.get("youtube_source_playlist_id") and session.get("spotify_destination_playlist_id"):
-        return True
-    if session.get("spotify_source_playlist_id") and session.get("youtube_destination_playlist_id"):
-        return True
+# TODO: Change this to POST
+@app.route("/startSync", methods=["GET"])
+def start_sync():
+    if not (session.get("youtube_source_playlist_id") and session.get("spotify_destination_playlist_id")):
+        return "Not supported yet"
 
-    return False
+    video_names = get_video_names_from_playlist(session.get("youtube_source_playlist_id"))
+    results = spotify_search_for_songs(video_names)
+    
+    add_tracks_to_spotify_playlist(results, session.get("spotify_destination_playlist_id"))
+
+    return str(json.dumps(results))
